@@ -3,7 +3,7 @@ import Flutter
 import flutter_downloader
 import Firebase
 
-@UIApplicationMain
+@main
 @objc class AppDelegate: FlutterAppDelegate {
   override func application(
     _ application: UIApplication,
