@@ -39,7 +39,7 @@ class _CourseAnnouncementPageState extends State<CourseAnnouncementPage>
 
   Color getColor(int index) {
     return (index % 2 == 1)
-        ? Theme.of(context).backgroundColor
+        ? Theme.of(context).scaffoldBackgroundColor
         : Theme.of(context).dividerColor;
   }
 
