@@ -63,7 +63,7 @@ class _CourseFolderPageState extends State<CourseFolderPage> {
 
   Color getColor(int index) {
     return (index % 2 == 1)
-        ? Theme.of(context).backgroundColor
+        ? Theme.of(context).scaffoldBackgroundColor
         : Theme.of(context).dividerColor;
   }
 
